FIX mismatching doctype
The document type is both declared as HTML5 (with the `<!doctype html>`
magic comment) and as XHTML (with the `xmlns="http://www.w3.org/1999/xhtml"`
attribute on the `<html>` element).

On most browsers, this is fine, but it causes browsers based on
webkit-gtk to prioritize XHTML, and then the page is not XHTML valid and
renders only partially, with errors.

I realized this because I use a webkit-gtk based browser, but it's
possible it affects webkit browsers as well, like Safari.

Anyway, the fix is easy : XHTML is a thing of the past. :)