Fixed anomalous planets messing with orbit numbers of main worlds in continuation method
Anomalous planet injection was changing the indexing of what orbit numbers were what orbit indices. I've added code to ensure when an anomalous planet is injected at an early orbit index than the main world's planned orbit index, the main world's planned orbit index is incremented. A note is also added to the log when this occurs.

This was causing the main world to occasionally be much hotter than it otherwise was supposed to be (as if you were going to make Earth the main world, but injected an anomalous planet between Venus and Mercury; this would cause Earth's UWP to be assigned to Venus's orbit instead, which would not be pleasant for the people on Earth).